# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FetchSheetRequest {
  /** The Google Sheets URL (regular share URL or CSV export URL) */
  sheetUrl: string;
  /** Optionally pass the sheet/tab ID (gid parameter) */
  sheetId?: string;
}

/**
 * A single row from the sheet, keyed by column header
 */
export interface SheetRow {
  [key: string]: string;
}

export type ColumnInfoType =
  (typeof ColumnInfoType)[keyof typeof ColumnInfoType];

export const ColumnInfoType = {
  string: "string",
  number: "number",
  date: "date",
  boolean: "boolean",
  unknown: "unknown",
} as const;

export interface ColumnInfo {
  name: string;
  type: ColumnInfoType;
  sampleValues: string[];
  nonEmptyCount: number;
}

export interface DetectedFeatures {
  hasCallMetrics: boolean;
  hasSentiment: boolean;
  hasCallDirection: boolean;
  hasCallStatus: boolean;
  hasTranscript: boolean;
  hasDriverInfo: boolean;
  hasBulkCalls: boolean;
  hasJoinInterest: boolean;
  hasRecording: boolean;
  hasPhoneNumbers: boolean;
}

export interface SheetDataResponse {
  rows: SheetRow[];
  columns: ColumnInfo[];
  totalRows: number;
  detectedFeatures: DetectedFeatures;
}

export interface SentimentBreakdown {
  positive: number;
  negative: number;
  neutral: number;
  unknown: number;
}

export interface CallDirectionBreakdown {
  inbound: number;
  outbound: number;
}

export interface CallStatusBreakdown {
  [key: string]: number;
}

export type SheetSummaryResponseTopBulkCallNamesItem = {
  name: string;
  count: number;
};

export interface SheetSummaryResponse {
  totalCalls: number;
  totalRows: number;
  avgDurationMinutes?: number;
  totalDurationMinutes?: number;
  sentimentBreakdown?: SentimentBreakdown;
  callDirectionBreakdown?: CallDirectionBreakdown;
  callStatusBreakdown?: CallStatusBreakdown;
  joinInterestCount?: number;
  topBulkCallNames?: SheetSummaryResponseTopBulkCallNamesItem[];
  detectedFeatures: DetectedFeatures;
  columns: ColumnInfo[];
}

export interface ErrorResponse {
  detail: string;
}

export interface ValidationError {
  loc: (string | number)[];
  msg: string;
  type: string;
}

export interface HTTPValidationError {
  detail: ValidationError[];
}