# packages/api-zod/src/generated/types/columnInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfoType } from "./columnInfoType";

export interface ColumnInfo {
  name: string;
  type: ColumnInfoType;
  sampleValues: string[];
  nonEmptyCount: number;
}