# packages/api-zod/src/generated/types/detectedFeatures.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DetectedFeatures {
  hasCallMetrics: boolean;
  hasSentiment: boolean;
  hasCallDirection: boolean;
  hasCallStatus: boolean;
  hasTranscript: boolean;
  hasDriverInfo: boolean;
  hasBulkCalls: boolean;
  hasJoinInterest: boolean;
  hasRecording: boolean;
  hasPhoneNumbers: boolean;
}