# packages/api-zod/src/generated/types/sheetDataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfo } from "./columnInfo";
import type { DetectedFeatures } from "./detectedFeatures";
import type { SheetRow } from "./sheetRow";

export interface SheetDataResponse {
  rows: SheetRow[];
  columns: ColumnInfo[];
  totalRows: number;
  detectedFeatures: DetectedFeatures;
}